+++ slipp/src/main/java/slipp/controller/ListUserController.java
@@ -1,6 +1,5 @@
 package slipp.controller;
 
-import nextstep.mvc.tobe.view.JspView;
 import nextstep.mvc.tobe.view.ModelAndView;
 import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
@@ -17,10 +16,10 @@ public class ListUserController {
     @RequestMapping(value = "/users", method = RequestMethod.GET)
     public ModelAndView showUsers(HttpServletRequest req, HttpServletResponse resp) {
         if (!UserSessionUtils.isLogined(req.getSession())) {
-            return new ModelAndView(new JspView("redirect:/users/loginForm"));
+            return new ModelAndView("redirect:/users/loginForm");
         }
 
         req.setAttribute("users", DataBase.findAll());
-        return new ModelAndView(new JspView("/user/list.jsp"));
+        return new ModelAndView("/user/list.jsp");
     }
 }
