+++ slipp/src/main/java/slipp/controller/ProfileController.java
@@ -1,6 +1,5 @@
 package slipp.controller;
 
-import nextstep.mvc.tobe.view.JspView;
 import nextstep.mvc.tobe.view.ModelAndView;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
@@ -20,6 +19,6 @@ public class ProfileController {
             throw new NullPointerException("사용자를 찾을 수 없습니다.");
         }
         req.setAttribute("user", user);
-        return new ModelAndView(new JspView("/user/profile.jsp"));
+        return new ModelAndView("/user/profile.jsp");
     }
 }
